2020.5.23

+++ hello-spring-boot/src/main/java/com/jemmy/hello/spring/boot/service/impl/UserService.java
@@ -24,6 +24,7 @@ public class UserService implements IUserService {
 
     /**
      * 登录逻辑
+     *
      * @param username
      * @param password
      * @return
@@ -32,12 +33,12 @@ public class UserService implements IUserService {
     public ServerResponse loginLogic(String username, String password) {
 
         //step1.用户名和密码的非空判断
-        if(null == username || username.equals("")){
+        if (null == username || username.equals("")) {
             //
             return ServerResponse.createServerResponseByFail(ResponseCode.USERNAME_EMPTY.getCode(),
                     ResponseCode.USERNAME_EMPTY.getMsg());
         }
-        if(null == password || password.equals("")){
+        if (null == password || password.equals("")) {
             //
             return ServerResponse.createServerResponseByFail(ResponseCode.PASSWORD_EMPTY.getCode(),
                     ResponseCode.PASSWORD_EMPTY.getMsg());
@@ -45,7 +46,7 @@ public class UserService implements IUserService {
 
         //step2.查看用户名是否存在
         Integer i = userMapper.findByUsername(username);
-        if(i==null || i.intValue()==0){
+        if (i == null || i == 0) {
             //用户名不存在
             return ServerResponse.createServerResponseByFail(ResponseCode.USERNAME_NOT_EXISTS.getCode(),
                     ResponseCode.USERNAME_NOT_EXISTS.getMsg());
@@ -53,7 +54,7 @@ public class UserService implements IUserService {
 
         //step3.根据用户名和密码查询
         User user = userMapper.findByUsernameAndPassword(username, password);
-        if(user == null){
+        if (user == null) {
             //密码错误
             return ServerResponse.createServerResponseByFail(ResponseCode.PASSWORD_ERROR.getCode(),
                     ResponseCode.PASSWORD_ERROR.getMsg());
@@ -62,22 +63,21 @@ public class UserService implements IUserService {
         //step4.返回结果
         //LoginUser增加
         int signOut = loginUserMapper.allSignOut();
-        if(signOut < 0){
+        if (signOut < 0) {
             return ServerResponse.createServerResponseByFail(ResponseCode.LOGIN_FAIL.getCode(),
                     ResponseCode.LOGIN_FAIL.getMsg());
         }
         LoginUser loginUser = loginUserMapper.selectByUsername(username);
-        if(loginUser == null){
+        if (loginUser == null) {
             int insert = loginUserMapper.putLoginUser(username);
-            if(insert == 0){
+            if (insert == 0) {
                 return ServerResponse.createServerResponseByFail(ResponseCode.LOGIN_FAIL.getCode(),
                         ResponseCode.LOGIN_FAIL.getMsg());
             }
-        }
-        else {
+        } else {
             loginUser.setStatus(0);
             int up = loginUserMapper.updateByPrimaryKey(loginUser);
-            if(up == 0){
+            if (up == 0) {
                 return ServerResponse.createServerResponseByFail(ResponseCode.LOGIN_FAIL.getCode(),
                         ResponseCode.LOGIN_FAIL.getMsg());
             }
@@ -87,12 +87,13 @@ public class UserService implements IUserService {
 
     /**
      * 注册逻辑
+     *
      * @param user
      * @return
      */
     @Override
     public ServerResponse registerLogic(User user) {
-        if(user == null){
+        if (user == null) {
             return ServerResponse.createServerResponseByFail(ResponseCode.PARAMETER_EMPTY.getCode(),
                     ResponseCode.PARAMETER_EMPTY.getMsg());
         }
@@ -103,17 +104,17 @@ public class UserService implements IUserService {
         user.setRole(Const.NORMAL_USER);
 
         //1.判断参数是否为空
-        if(StringUtils.isBlank(username)){
+        if (StringUtils.isBlank(username)) {
             //
             return ServerResponse.createServerResponseByFail(ResponseCode.USERNAME_EMPTY.getCode(),
                     ResponseCode.USERNAME_EMPTY.getMsg());
         }
-        if(null == password || password.equals("")){
+        if (null == password || password.equals("")) {
             //
             return ServerResponse.createServerResponseByFail(ResponseCode.PASSWORD_EMPTY.getCode(),
                     ResponseCode.PASSWORD_EMPTY.getMsg());
         }
-        if(null == phone || phone.equals("")){
+        if (null == phone || phone.equals("")) {
             //
             return ServerResponse.createServerResponseByFail(ResponseCode.PHONE_EMPTY.getCode(),
                     ResponseCode.PHONE_EMPTY.getMsg());
@@ -121,21 +122,21 @@ public class UserService implements IUserService {
 
         //2.判断用户名是否存在
         Integer i = userMapper.findByUsername(username);
-        if(i != null && i > 0){//用户名存在
+        if (i != null && i > 0) {//用户名存在
             return ServerResponse.createServerResponseByFail(ResponseCode.USERNAME_EXISTS.getCode(),
                     ResponseCode.USERNAME_EXISTS.getMsg());
         }
 
         //3.判断手机号是否存在
         Integer i2 = userMapper.findByPhone(phone);
-        if(i2 != null && i2 > 0){//手机存在
+        if (i2 != null && i2 > 0) {//手机存在
             return ServerResponse.createServerResponseByFail(ResponseCode.PHONE_EXISTS.getCode(),
                     ResponseCode.PHONE_EXISTS.getMsg());
         }
 
         //4.注册  (加密？)
         int result = userMapper.insert(user);
-        if(result == 0){
+        if (result == 0) {
             return ServerResponse.createServerResponseByFail(ResponseCode.REGISTER_FAIL.getCode(),
                     ResponseCode.REGISTER_FAIL.getMsg());
         }
@@ -152,13 +153,14 @@ public class UserService implements IUserService {
 
     /**
      * 注销
+     *
      * @param username
      * @return
      */
     @Override
-    public ServerResponse signOut(String username){
+    public ServerResponse signOut(String username) {
         int result = loginUserMapper.signOutByUsername(username);
-        if(result == 0){
+        if (result == 0) {
             return ServerResponse.createServerResponseByFail(ResponseCode.SIGN_OUT_FAIL.getCode(),
                     ResponseCode.SIGN_OUT_FAIL.getMsg());
         }
@@ -169,7 +171,7 @@ public class UserService implements IUserService {
     public ServerResponse changepwdLogic(Integer id, String old, String now) {
 
         //判空
-        if(StringUtils.isBlank(old) || StringUtils.isBlank(now)){
+        if (StringUtils.isBlank(old) || StringUtils.isBlank(now)) {
             return ServerResponse.createServerResponseByFail(ResponseCode.PARAMETER_EMPTY.getCode(),
                     ResponseCode.PARAMETER_EMPTY.getMsg());
         }
@@ -180,20 +182,20 @@ public class UserService implements IUserService {
 
         //若原密码不对
         boolean same = userMapper.selectByPrimaryKey(id).getPassword().equals(old);
-        if(!same){
+        if (!same) {
             return ServerResponse.createServerResponseByFail(ResponseCode.PASSWORD_ERROR.getCode(),
                     ResponseCode.PASSWORD_ERROR.getMsg());
         }
 
         //若新旧密码一样
-        if(old.equals(now)){
+        if (old.equals(now)) {
             return ServerResponse.createServerResponseByFail(ResponseCode.PASSWORD_IS_SAME.getCode(),
                     ResponseCode.PASSWORD_IS_SAME.getMsg());
         }
 
         //改密码
         int result = userMapper.changePasswordById(id, now);
-        if(result == 0){  //改不了？
+        if (result == 0) {  //改不了？
             return ServerResponse.createServerResponseByFail(ResponseCode.CHANGE_PASSWORD_FAIL.getCode(),
                     ResponseCode.CHANGE_PASSWORD_FAIL.getMsg());
         }
